+++ OnlineOrderManagementSystem.Repository/IServices/Sal/IProductService.cs
@@ -12,8 +12,8 @@ namespace OnlineOrderManagementSystem.Repository.IServices.Sal
     {
         Task<bool> AddProductAsync(AddProductDTO product);
         Task<bool> UpdateProductAsync(Product product);
-        Task<bool> DeleteProductAsync(int productId);
-        Task<Product?> GetProductByIdAsync(int productId);
+        Task<bool> DeleteProductAsync(long productId);
+        Task<Product?> GetProductByIdAsync(long productId);
         Task<IEnumerable<Product>> GetAllProductsAsync(GetProductsDTO dto);
     }
 }
